+++ packages/ui/src/__mocks__/logto.tsx
@@ -1,133 +1,20 @@
 import type { SignInExperience, SignIn } from '@logto/schemas';
-import { ConnectorPlatform, ConnectorType, SignInIdentifier, SignInMode } from '@logto/schemas';
+import {
+  ConnectorPlatform,
+  ConnectorType,
+  MfaPolicy,
+  SignInIdentifier,
+  SignInMode,
+} from '@logto/schemas';
 
 import type { SignInExperienceResponse } from '@/types';
 
+import { socialConnectors } from './social-connectors';
+
+export * from './social-connectors';
+
 export const appLogo = 'https://avatars.githubusercontent.com/u/88327661?s=200&v=4';
 export const appHeadline = 'Build user identity in a modern way';
-
-export const socialConnectors = [
-  {
-    id: 'BE8QXN0VsrOH7xdWFDJZ9',
-    target: 'github',
-    platform: ConnectorPlatform.Web,
-    type: ConnectorType.Social,
-    logo: 'https://user-images.githubusercontent.com/5717882/156983224-7ea0296b-38fa-419d-9515-67e8a9612e09.png',
-    logoDark: null,
-    name: {
-      en: 'Sign in with GitHub',
-      'pt-PT': 'Entrar com GitHub',
-      'zh-CN': '使用 GitHub 登录',
-      'tr-TR': 'Github ile giriş yap',
-      ko: 'Github 로그인',
-    },
-    description: {
-      en: 'Sign in with GitHub',
-      'pt-PT': 'Entrar com GitHub',
-      'zh-CN': '使用 GitHub 登录',
-      'tr-TR': 'Github ile giriş yap',
-      ko: 'Github 로그인',
-    },
-    readme: '',
-    configTemplate: '',
-  },
-  {
-    id: '24yt_xIUl5btN4UwvFokt',
-    target: 'alipay',
-    platform: ConnectorPlatform.Web,
-    type: ConnectorType.Social,
-    logo: 'https://user-images.githubusercontent.com/5717882/156983224-7ea0296b-38fa-419d-9515-67e8a9612e09.png',
-    logoDark: null,
-    name: {
-      en: 'Sign in with Alipay',
-      'pt-PT': 'Entrar com Alipay',
-      'zh-CN': '使用 Alipay 登录',
-      'tr-TR': 'Alipay ile giriş yap',
-      ko: 'Alipay 로그인',
-    },
-    description: {
-      en: 'Sign in with Alipay',
-      'pt-PT': 'Entrar com Alipay',
-      'zh-CN': '使用 Alipay 登录',
-      'tr-TR': 'Alipay ile giriş yap',
-      ko: 'Alipay 로그인',
-    },
-    readme: '',
-    configTemplate: '',
-  },
-  {
-    id: 'E5kb2gdq769qOEYaLg1V5',
-    target: 'wechat',
-    platform: ConnectorPlatform.Web,
-    type: ConnectorType.Social,
-    logo: 'https://user-images.githubusercontent.com/5717882/156983224-7ea0296b-38fa-419d-9515-67e8a9612e09.png',
-    logoDark: null,
-    name: {
-      en: 'Sign in with WeChat',
-      'pt-PT': 'Entrar com WeChat',
-      'zh-CN': '使用 WeChat 登录',
-      'tr-TR': 'WeChat ile giriş yap',
-      ko: 'WeChat 로그인',
-    },
-    description: {
-      en: 'Sign in with WeChat',
-      'pt-PT': 'Entrar com WeChat',
-      'zh-CN': '使用 WeChat 登录',
-      'tr-TR': 'WeChat ile giriş yap',
-      ko: 'WeChat 로그인',
-    },
-    readme: '',
-    configTemplate: '',
-  },
-  {
-    id: 'xY2YZEweMFPKxphngGHhy',
-    target: 'google',
-    platform: ConnectorPlatform.Web,
-    type: ConnectorType.Social,
-    logo: 'https://user-images.githubusercontent.com/5717882/156983224-7ea0296b-38fa-419d-9515-67e8a9612e09.png',
-    logoDark: null,
-    name: {
-      en: 'Sign in with Google',
-      'pt-PT': 'Entrar com Google',
-      'zh-CN': '使用 Google 登录',
-      'tr-TR': 'Google ile giriş yap',
-      ko: 'Google 로그인',
-    },
-    description: {
-      en: 'Sign in with Google',
-      'pt-PT': 'Entrar com Google',
-      'zh-CN': '使用 Google 登录',
-      'tr-TR': 'Google ile giriş yap',
-      ko: 'Google 로그인',
-    },
-    readme: '',
-    configTemplate: '',
-  },
-  {
-    id: 'lcXT4o2GSjbV9kg2shZC7',
-    target: 'facebook',
-    platform: ConnectorPlatform.Web,
-    type: ConnectorType.Social,
-    logo: 'https://user-images.githubusercontent.com/5717882/156983224-7ea0296b-38fa-419d-9515-67e8a9612e09.png',
-    logoDark: null,
-    name: {
-      en: 'Sign in with Meta',
-      'pt-PT': 'Entrar com Facebook',
-      'zh-CN': '使用 Meta 登录',
-      'tr-TR': 'Meta ile giriş yap',
-      ko: 'Meta 로그인',
-    },
-    description: {
-      en: 'Sign in with Meta',
-      'pt-PT': 'Entrar com Facebook',
-      'zh-CN': '使用 Meta 登录',
-      'tr-TR': 'Meta ile giriş yap',
-      ko: 'Meta 로그인',
-    },
-    readme: '',
-    configTemplate: '',
-  },
-];
 
 export const mockSocialConnectorData = {
   id: 'arbitrary-social-connector-data',
@@ -205,6 +92,10 @@ export const mockSignInExperience: SignInExperience = {
   customCss: null,
   customContent: {},
   passwordPolicy: {},
+  mfa: {
+    policy: MfaPolicy.UserControlled,
+    factors: [],
+  },
 };
 
 export const mockSignInExperienceSettings: SignInExperienceResponse = {
@@ -230,6 +121,10 @@ export const mockSignInExperienceSettings: SignInExperienceResponse = {
   customCss: null,
   customContent: {},
   passwordPolicy: {},
+  mfa: {
+    policy: MfaPolicy.UserControlled,
+    factors: [],
+  },
 };
 
 const usernameSettings = {
